Fix bug. Change datatype from int to bigint in serial number.

+++ db/testserver_sp_dcct.sql
@@ -6,7 +6,7 @@ delimiter $$
 use testserverdb$$
 create procedure sp_add_dcct
 (
-		in p_numero_serie					int
+		in p_numero_serie					bigint(15)
 )
 
 begin
@@ -31,58 +31,58 @@ delimiter $$
 use testserverdb$$
 create procedure sp_add_log_dcct
 (
-    in p_resultado_teste						varchar(45),
-		in p_numero_serie_dcct       		int,
-		in p_iload0											double,
-		in p_iload1											double,
-		in p_iload2											double,
-		in p_iload3											double,
-		in p_iload4											double,
-		in p_iload5											double,
-		in p_iload6											double,
-		in p_iload7											double,
-		in p_iload8											double,
-		in p_iload9											double,
-		in p_iload10										double,
-		in p_details										text
+    in p_resultado_teste		varchar(45),
+	in p_numero_serie_dcct      bigint(15),
+	in p_iload0					double,
+	in p_iload1					double,
+	in p_iload2					double,
+	in p_iload3					double,
+	in p_iload4					double,
+	in p_iload5					double,
+	in p_iload6					double,
+	in p_iload7					double,
+	in p_iload8					double,
+	in p_iload9					double,
+	in p_iload10				double,
+	in p_details				text
 )
 
 begin
 insert into LogDcct
 (
-		data,
+	data,
     resultado_teste,
     numero_serie_dcct,
-		iload0,
-		iload1,
-		iload2,
-		iload3,
-		iload4,
-		iload5,
-		iload6,
-		iload7,
-		iload8,
-		iload9,
-		iload10,
-		details
+	iload0,
+	iload1,
+	iload2,
+	iload3,
+	iload4,
+	iload5,
+	iload6,
+	iload7,
+	iload8,
+	iload9,
+	iload10,
+	details
 )
 values
 (
     now(),
     p_resultado_teste,
     p_numero_serie_dcct,
-		p_iload0,
-		p_iload1,
-		p_iload2,
-		p_iload3,
-		p_iload4,
-		p_iload5,
-		p_iload6,
-		p_iload7,
-		p_iload8,
-		p_iload9,
-		p_iload10,
-		p_details
+	p_iload0,
+	p_iload1,
+	p_iload2,
+	p_iload3,
+	p_iload4,
+	p_iload5,
+	p_iload6,
+	p_iload7,
+	p_iload8,
+	p_iload9,
+	p_iload10,
+	p_details
 );
 end$$
 delimiter ;
